chore(vue-cli-plugin-eslint): Add missing translations, extract UI descriptors
