+++ packages/@vue/cli-plugin-eslint/ui/index.js
@@ -1,0 +1,40 @@
+const configDescriptor = require('./configDescriptor')
+const taskDescriptor = require('./taskDescriptor')
+
+const CONFIG = 'org.vue.eslintrc'
+const OPEN_ESLINTRC = 'org.vue.eslint.open-eslintrc'
+
+module.exports = api => {
+  api.describeConfig(configDescriptor)
+  api.describeTask(taskDescriptor)
+
+  api.onViewOpen(({ view }) => {
+    if (view.id !== 'vue-project-configurations') {
+      removeSuggestions()
+    }
+  })
+
+  api.onConfigRead(({ config }) => {
+    if (config.id === CONFIG) {
+      api.addSuggestion({
+        id: OPEN_ESLINTRC,
+        type: 'action',
+        label: 'org.vue.eslint.suggestions.open-eslintrc.label',
+        handler () {
+          const file = config.foundFiles.eslint.path
+          const launch = require('launch-editor')
+          launch(file)
+          return {
+            keep: true
+          }
+        }
+      })
+    } else {
+      removeSuggestions()
+    }
+  })
+
+  function removeSuggestions () {
+    [OPEN_ESLINTRC].forEach(id => api.removeSuggestion(id))
+  }
+}
