+++ packages/@vue/cli-plugin-eslint/ui/taskDescriptor.js
@@ -1,0 +1,16 @@
+module.exports = {
+  match: /vue-cli-service lint/,
+  description: 'org.vue.eslint.tasks.lint.description',
+  link: 'https://github.com/vuejs/vue-cli/tree/dev/packages/%40vue/cli-plugin-eslint#injected-commands',
+  prompts: [
+    {
+      name: 'noFix',
+      type: 'confirm',
+      default: false,
+      description: 'org.vue.eslint.tasks.lint.noFix'
+    }
+  ],
+  onBeforeRun: ({ answers, args }) => {
+    if (answers.noFix) args.push('--no-fix')
+  }
+}
